requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
pytest==5.1.1
selenium==3.14.0

